Splits MatchFinder
