+++ Jantimizer/QueryOptimiser/Cost/EstimateCalculators/EstimateCalculatorEquiDepth.cs
@@ -2,6 +2,7 @@
 using DatabaseConnector;
 using Histograms;
 using Histograms.Models;
+using QueryOptimiser.Cost.EstimateCalculators.MatchFinders;
 using QueryOptimiser.Cost.Nodes;
 using QueryOptimiser.Models;
 using QueryParser.Models;
@@ -10,14 +11,12 @@ namespace QueryOptimiser.Cost.EstimateCalculators
 {
     internal class EstimateCalculatorEquiDepth : BaseEstimateCalculator
     {
-        public override IJoinEstimate JoinEstimator { get; set; }
-        public override IFilterEstimate FilterEstimator { get; set; }
-        public override MatchFinder Matcher { get; set; }
+        public override JoinMatchFinder JoinMatcher { get; set; }
+        public override FilterMatchFinder FilterMatcher { get; set; }
 
-        public EstimateCalculatorEquiDepth(IHistogramManager manager) : base(manager) {
-            JoinEstimator = new JoinEstimateEquiDepth();
-            FilterEstimator = new FilterEstimateEquiDepth();
-            Matcher = new MatchFinder(JoinEstimator, FilterEstimator);
+        internal EstimateCalculatorEquiDepth(IHistogramManager manager) : base(manager) {
+            JoinMatcher = new JoinMatchFinder(new JoinEstimateEquiDepth());
+            FilterMatcher = new FilterMatchFinder(new FilterEstimateEquiDepth());
         }
     }
 }
