+++ Jantimizer/QueryOptimiser/Cost/EstimateCalculators/IEstimateCalculator.cs
@@ -10,14 +10,15 @@ using Histograms.Models;
 using DatabaseConnector;
 using QueryOptimiser.Models;
 using QueryOptimiser.Cost.Nodes;
+using QueryOptimiser.Cost.EstimateCalculators.MatchFinders;
 
 namespace QueryOptimiser.Cost.EstimateCalculators
 {
     public interface IEstimateCalculator
     {
-        public IHistogramManager HistogramManager { get; }
-        public IJoinEstimate JoinEstimator { get; }
-        public IFilterEstimate FilterEstimator { get; }
+        public IHistogramManager HistogramManager { get; set; }
+        public JoinMatchFinder JoinMatcher { get; set; }
+        public FilterMatchFinder FilterMatcher { get; set; }
 
         public IntermediateTable EstimateIntermediateTable(INode node, IntermediateTable intermediateTable);
     }
