+++ Jantimizer/QueryOptimiser/Cost/EstimateCalculators/MatchFinders/IMatchFinder.cs
@@ -1,0 +1,15 @@
+﻿using QueryOptimiser.Cost.Nodes;
+using QueryParser.Models;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace QueryOptimiser.Cost.EstimateCalculators.MatchFinders
+{
+    public interface IMatchFinder<EstimatorType>
+        where EstimatorType : INodeEstimate
+    {
+    }
+}
