+++ Jantimizer/QueryOptimiserTest/Cost/Nodes/JoinEstimateEquiDepthVarianceTests.cs
@@ -42,90 +42,5 @@ namespace QueryOptimiserTest.Cost.Nodes
         }
 
         #endregion
-
-        #region GetBucketCertainty
-
-        [TestMethod]
-        [DataRow(1, 1, 1)]
-        [DataRow(1, 2, 0.5)]
-        [DataRow(1, 10, 0.1)]
-        [DataRow(100, 10, 10)]
-        public void Can_GetBucketCertainty(double stdDeviation, double range, double expResult)
-        {
-            // ARRANGE
-            HistogramBucketVariance bucket = new HistogramBucketVariance(0, 1, 1, 1, 1, stdDeviation, range);
-
-            var joinCost = new JoinEstimateEquiDepthVariance();
-
-            // ACT
-            double bCertainty = joinCost.GetBucketCertainty(bucket);
-
-            // ASSERTs
-            Assert.AreEqual(expResult, bCertainty);
-        }
-
-        [TestMethod]
-        public void Can_GetBucketCertainty_If_Bucket_RangeZero()
-        {
-            // ARRANGE
-            HistogramBucketVariance bucket = new HistogramBucketVariance(0, 1, 1, 1, 1, 1, 0);
-
-            var joinCost = new JoinEstimateEquiDepthVariance();
-
-            // ACT
-            double bCertainty = joinCost.GetBucketCertainty(bucket);
-
-            // ASSERTs
-            Assert.AreEqual(1, bCertainty);
-        }
-
-        [TestMethod]
-        public void Can_GetBucketCertainty_If_Bucket_StandardDeviationZero()
-        {
-            // ARRANGE
-            HistogramBucketVariance bucket = new HistogramBucketVariance(0, 1, 1, 1, 1, 0, 1);
-
-            var joinCost = new JoinEstimateEquiDepthVariance();
-
-            // ACT
-            double bCertainty = joinCost.GetBucketCertainty(bucket);
-
-            // ASSERTs
-            Assert.AreEqual(1, bCertainty);
-        }
-
-        #endregion
-
-        #region GetTotalCertainty
-
-        [TestMethod]
-        [DataRow(1, 1, 1)]
-        [DataRow(1, 2, 0.5)]
-        [DataRow(2, 1, 0.5)]
-        [DataRow(100, 1, 0.01)]
-        public void Can_GetTotalCertainty(double bucketCertainty, double comparisonBucketCertainty, double expResult)
-        {
-            // ARRANGE
-            var joinCost = new JoinEstimateEquiDepthVariance();
-
-            // ACT
-            double certainty = joinCost.GetTotalCertainty(bucketCertainty, comparisonBucketCertainty);
-
-            // ASSERTs
-            Assert.AreEqual(expResult, certainty);
-        }
-
-        [TestMethod]
-        [ExpectedException(typeof(ArgumentOutOfRangeException))]
-        public void Cant_GetTotalCertainty_IfComparisonIsZero()
-        {
-            // ARRANGE
-            var joinCost = new JoinEstimateEquiDepthVariance();
-
-            // ACT
-            joinCost.GetTotalCertainty(1, 0);
-        }
-
-        #endregion
     }
 }
