+++ Jantimizer/QueryParserTest/ParserManagerTests.cs
@@ -39,7 +39,7 @@ namespace QueryParserTest
             // ARRANGE
             TestParser parser = new TestParser();
             parser.ShouldAccept = true;
-            parser.ShouldReturn = new List<INode>() { new TestNode(0), new TestNode(1) };
+            parser.ShouldReturn = new ParserResult() { };
             IParserManager newManager = new ParserManager(new List<IQueryParser>() { parser });
 
             // ACT
@@ -56,9 +56,9 @@ namespace QueryParserTest
             TestParser parser1 = new TestParser();
             TestParser parser2 = new TestParser();
             parser1.ShouldAccept = true;
-            parser1.ShouldReturn = new List<INode>() { new TestNode(0), new TestNode(1) };
+            parser1.ShouldReturn = new ParserResult() { };
             parser2.ShouldAccept = false;
-            parser2.ShouldReturn = new List<INode>() { new TestNode(2), new TestNode(3) };
+            parser2.ShouldReturn = new ParserResult() { };
             IParserManager newManager = new ParserManager(new List<IQueryParser>() { parser2, parser1 });
 
             // ACT
@@ -78,7 +78,7 @@ namespace QueryParserTest
             var result = newManager.ParseQuery("anything", false);
 
             // ASSERT
-            Assert.IsTrue(result.Count == 0);
+            Assert.IsTrue(result.GetNodes().Count == 0);
         }
 
         [TestMethod]
@@ -201,7 +201,7 @@ namespace QueryParserTest
     internal class TestParser : IQueryParser
     {
         public bool ShouldAccept = false;
-        public List<INode> ShouldReturn = new List<INode>();
+        public ParserResult ShouldReturn = new ParserResult();
 
         public bool DoesQueryMatch(string query)
         {
@@ -213,12 +213,12 @@ namespace QueryParserTest
             return Task.Run(() => ShouldAccept);
         }
 
-        public List<INode> ParseQuery(string query)
+        public ParserResult ParseQuery(string query)
         {
             return ShouldReturn;
         }
 
-        public Task<List<INode>> ParseQueryAsync(string query)
+        public Task<ParserResult> ParseQueryAsync(string query)
         {
             return Task.Run(() => ShouldReturn);
         }
